moved to avx cpu
